requirements.txt
Flask
pymongo[srv]
Flask-PyMongo
Flask-Cors
python-dotenv
